+++ projects/packages/jetpack-mu-wpcom/src/features/launchpad/launchpad.php
@@ -479,11 +479,11 @@ function wpcom_hacky_track_video_uploaded_task( $post_id ) {
 	if ( get_option( 'launchpad_screen' ) !== 'full' ) {
 		return;
 	}
-	if ( has_action( 'add_attachment', 'wpcom_track_video_uploaded_task' ) ) {
+	if ( has_action( 'add_attachment', 'wpcom_launchpad_track_video_uploaded_task' ) ) {
 		return;
 	}
 
-	wpcom_track_video_uploaded_task( $post_id );
+	wpcom_launchpad_track_video_uploaded_task( $post_id );
 }
 add_action( 'add_attachment', 'wpcom_hacky_track_video_uploaded_task' );
 
@@ -686,14 +686,10 @@ function wpcom_launchpad_set_fse_next_steps_modal_hidden( $should_hide ) {
  * @return bool True if the task list is enabled, false otherwise.
  */
 function wpcom_launchpad_is_keep_building_enabled() {
-	$intent                  = get_option( 'site_intent', false );
-	$launchpad_task_statuses = get_option( 'launchpad_checklist_tasks_statuses', array() );
+	$intent  = get_option( 'site_intent', false );
+	$blog_id = get_current_blog_id();
 
-	// We don't care about the other *_launched tasks, since this is specific to the Build flow.
-	$launched = isset( $launchpad_task_statuses['site_launched'] ) && $launchpad_task_statuses['site_launched'];
-	$blog_id  = get_current_blog_id();
-
-	if ( 'build' === $intent && $blog_id > 220443356 && $launched ) {
+	if ( 'build' === $intent && $blog_id > 220443356 ) {
 		return true;
 	}
 
@@ -720,7 +716,7 @@ function wpcom_launchpad_is_free_newsletter_enabled() {
 		return false;
 	}
 
-	return ! wpcom_has_goal_paid_subscribers() && apply_filters( 'wpcom_launchpad_intent_free_newsletter_enabled', false );
+	return ! wpcom_launchpad_has_goal_paid_subscribers() && apply_filters( 'wpcom_launchpad_intent_free_newsletter_enabled', false );
 }
 
 /**
@@ -734,7 +730,7 @@ function wpcom_launchpad_is_paid_newsletter_enabled() {
 		return false;
 	}
 
-	return wpcom_has_goal_paid_subscribers() && apply_filters( 'wpcom_launchpad_intent_paid_newsletter_enabled', false );
+	return wpcom_launchpad_has_goal_paid_subscribers() && apply_filters( 'wpcom_launchpad_intent_paid_newsletter_enabled', false );
 }
 
 // Unhook our old mu-plugin - this current file is being loaded on 0 priority for `plugins_loaded`.