+++ projects/packages/jetpack-mu-wpcom/src/features/wpcom-endpoints/class-wpcom-rest-api-v2-endpoint-launchpad.php
@@ -103,7 +103,10 @@ class WPCOM_REST_API_V2_Endpoint_Launchpad extends WP_REST_Controller {
 		$tasks            = wpcom_launchpad_checklists()->get_all_tasks();
 		$allowed_task_ids = array();
 		foreach ( $tasks as $task ) {
-			$allowed_task_ids[] = wpcom_launchpad_checklists()->get_task_key( $task );
+			$allowed_task_ids[] = $task['id'];
+			if ( isset( $task['id_map'] ) ) {
+				$allowed_task_ids[] = $task['id_map'];
+			}
 		}
 		$allowed_task_ids = array_unique( $allowed_task_ids );
 		$properties       = array();