+++ projects/packages/forms/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.21.0] - 2023-09-04
+### Added
+- Add Jetpack AI Form section to new Forms landing page [#32726]
+
+### Changed
+- Updated package dependencies. [#32803]
+- Updated package dependencies. [#32804]
+
+### Fixed
+- Fix block icons for display on wp.org [#32754]
+
 ## [0.20.1] - 2023-08-28
 ### Changed
 - Updated package dependencies. [#32605]
@@ -316,6 +327,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a new jetpack/forms package [#28409]
 - Added a public load_contact_form method for initializing the contact form module. [#28416]
 
+[0.21.0]: https://github.com/automattic/jetpack-forms/compare/v0.20.1...v0.21.0
 [0.20.1]: https://github.com/automattic/jetpack-forms/compare/v0.20.0...v0.20.1
 [0.20.0]: https://github.com/automattic/jetpack-forms/compare/v0.19.11...v0.20.0
 [0.19.11]: https://github.com/automattic/jetpack-forms/compare/v0.19.10...v0.19.11